refactor(connector): rename all files temporarily used for refactoring to the formal name. (#323)
* refactor(connector): rename all files temporarily used for refactor to the formal name.

* fix(connector): fix some conflict and test package name.
