--- packages/connector/interface/src/annotations/Dbt.kt
@@ -1,38 +1,0 @@
-package io.tellery.connectors.annotations
-
-import io.tellery.annotations.Config
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class Dbt(
-    val type: String,
-    val configs: Array<Config> = [
-        Config(
-            name = DbtFields.GIT_URL,
-            type = Config.ConfigType.STRING,
-            description = "The git url of your dbt repository. The format should be: git://[endpoint]:[organization_name]/[repo_name].git.",
-            hint = "",
-            required = false
-        ),
-        Config(
-            name = DbtFields.DBT_PROJECT_NAME,
-            type = Config.ConfigType.STRING,
-            description = "The name of your dbt project, it should be equal with the name in dbt_project.yml.",
-            hint = "",
-            required = false
-        ),
-        Config(
-            name = DbtFields.PUBLIC_KEY,
-            type = Config.ConfigType.STRING,
-            description = "The public key generated by tellery for dbt repository, please copy it to the 'deploy key' in git server.",
-            hint = "",
-            required = false
-        )
-    ]
-) {
-    object DbtFields {
-        const val GIT_URL = "Git Url"
-        const val DBT_PROJECT_NAME = "Dbt Project Name"
-        const val PUBLIC_KEY = "Public Key"
-    }
-}
