+++ packages/connector/interface/src/entities/Entity.kt
