+++ packages/connector/interface/src/integrations/DbtIntegration.kt
@@ -5,29 +5,31 @@ import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import io.grpc.Status
 import io.tellery.annotations.Config
 import io.tellery.annotations.Integration
-import io.tellery.connectors.annotations.Dbt
 import io.tellery.entities.CustomizedException
 import io.tellery.entities.ProfileEntity
+import io.tellery.integrations.DbtIntegration.Companion.DBT_PROJECT_FIELD
+import io.tellery.integrations.DbtIntegration.Companion.GIT_URL_FIELD
+import io.tellery.integrations.DbtIntegration.Companion.PUBLIC_KEY_FIELD
 
 @Integration(
     type = "dbt",
     configs = [
         Config(
-            name = Dbt.DbtFields.GIT_URL,
+            name = GIT_URL_FIELD,
             type = Config.ConfigType.STRING,
             description = "The git url of your dbt repository. The format should be: git://[endpoint]:[organization_name]/[repo_name].git.",
             hint = "",
             required = true
         ),
         Config(
-            name = Dbt.DbtFields.DBT_PROJECT_NAME,
+            name = DBT_PROJECT_FIELD,
             type = Config.ConfigType.STRING,
             description = "The name of your dbt project, it should be equal with the name in dbt_project.yml.",
             hint = "",
             required = true
         ),
         Config(
-            name = Dbt.DbtFields.PUBLIC_KEY,
+            name = PUBLIC_KEY_FIELD,
             type = Config.ConfigType.STRING,
             description = "The public key generated by tellery for dbt repository, please copy it to the 'deploy key' in git server.",
             hint = "",
@@ -38,6 +40,10 @@ import io.tellery.entities.ProfileEntity
 abstract class DbtIntegration : BaseIntegration() {
 
     companion object {
+        const val GIT_URL_FIELD = "Git Url"
+        const val PUBLIC_KEY_FIELD = "Public Key"
+        const val DBT_PROJECT_FIELD = "Dbt Project Name"
+
         val mapper = jacksonObjectMapper()
     }
