--- packages/connector/server/src/common/dbt/DbtProfileWrapper.kt
@@ -1,8 +1,0 @@
-package io.tellery.common.dbt
-
-import io.tellery.connectors.profiles.BaseDbtProfile
-
-
-data class Entity(val outputs: Output, val target: String = "dev") {
-    data class Output(val dev: BaseDbtProfile)
-}
