--- packages/connector/server/src/common/dbt/model/Manifest.kt
@@ -1,9 +1,0 @@
-package io.tellery.common.dbt.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class Manifest(
-    val nodes: Map<String, DbtModel>,
-    val sources: Map<String, DbtModel>
-)
