+++ packages/connector/server/src/utils/CommonUtil.kt
