+++ packages/connector/server/test/managers/DbtManagerTest.kt
@@ -1,4 +1,4 @@
-package managers
+package io.tellery.managers
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
@@ -20,9 +20,7 @@ import io.tellery.entities.IntegrationEntity
 import io.tellery.entities.ProfileEntity
 import io.tellery.grpc.DbtBlock
 import io.tellery.grpc.QuestionBlockContent
-import io.tellery.managers.DbtManagerV2
-import io.tellery.managers.ProfileManager
-import io.tellery.utils.GitUtilsV2
+import io.tellery.utils.GitUtil
 import org.apache.commons.io.FileUtils
 import java.io.BufferedReader
 import java.io.File
@@ -78,7 +76,7 @@ class DbtManagerTest : FunSpec({
                 )
             } returns integrationEntity
 
-            val dbtManager = DbtManagerV2(pm)
+            val dbtManager = DbtManager(pm)
             dbtManager.setDbtProfilePath(profilePath)
             dbtManager.reloadContext()
 
@@ -95,7 +93,7 @@ class DbtManagerTest : FunSpec({
         every { pm.getProfileById(workspaceId) } returns null
         every { pm.getIntegrationInProfileAndByType(workspaceId, "dbt") } returns null
 
-        val dbtManager = DbtManagerV2(pm)
+        val dbtManager = DbtManager(pm)
         dbtManager.setDbtProfilePath(profilePath)
         dbtManager.reloadContext()
 
@@ -114,9 +112,9 @@ class DbtManagerTest : FunSpec({
             nameFn = { t -> "${if (t.value == 1) "call" else "don't call"} commitAndPush" },
             mockMap.entries
         ) { (path, time) ->
-            mockkObject(GitUtilsV2)
-            every { GitUtilsV2.commitAndPush(any(), any(), any()) } returns mockk()
-            every { GitUtilsV2.checkoutMasterAndPull(any(), any()) } returns mockk()
+            mockkObject(GitUtil)
+            every { GitUtil.commitAndPush(any(), any(), any()) } returns mockk()
+            every { GitUtil.checkoutMasterAndPull(any(), any()) } returns mockk()
 
             every { pm.getProfileById(workspaceId) } returns
                     mapper.readValue<ProfileEntity>(getResourceFileURL("/profiles/profile_bigquery.json"))
@@ -128,7 +126,7 @@ class DbtManagerTest : FunSpec({
             } returns integrationEntity
 
             // mock env and data
-            val dbtManager = DbtManagerV2(pm)
+            val dbtManager = DbtManager(pm)
             dbtManager.reloadContext()
             val repoDir = dbtManager.getContext()?.repoDir!!
             FileUtils.forceMkdir(repoDir.toFile())
@@ -140,14 +138,14 @@ class DbtManagerTest : FunSpec({
 
             // assert
             projectFilePath.readText() shouldContain "ephemeral"
-            verify(exactly = time) { GitUtilsV2.commitAndPush(any(), any(), any()) }
+            verify(exactly = time) { GitUtil.commitAndPush(any(), any(), any()) }
         }
     }
 
     context("Should insert a block while the block was not there before.") {
-        mockkObject(GitUtilsV2)
-        every { GitUtilsV2.checkoutMasterAndPull(any(), any()) } returns mockk()
-        every { GitUtilsV2.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
+        mockkObject(GitUtil)
+        every { GitUtil.checkoutMasterAndPull(any(), any()) } returns mockk()
+        every { GitUtil.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
 
         every { pm.getProfileById(workspaceId) } returns
                 mapper.readValue<ProfileEntity>(getResourceFileURL("/profiles/profile_bigquery.json"))
@@ -155,7 +153,7 @@ class DbtManagerTest : FunSpec({
 
         // mock env and data
         val block = providesQuestionBlockContent("/dbt/question_block_latest.json")
-        val dbtManager = DbtManagerV2(pm)
+        val dbtManager = DbtManager(pm)
         dbtManager.reloadContext()
         val telleryDir = dbtManager.getContext()?.repoDir?.resolve("models/tellery")!!
         FileUtils.forceMkdir(telleryDir.toFile())
@@ -166,7 +164,7 @@ class DbtManagerTest : FunSpec({
         // assert
         telleryDir.listDirectoryEntries().map { it.name } shouldContain "${block.name}.sql"
         verify(exactly = 1) {
-            GitUtilsV2.checkoutNewBranchAndCommitAndPush(
+            GitUtil.checkoutNewBranchAndCommitAndPush(
                 any(),
                 any(),
                 "feat(tellery): async model files in tellery dir.\ninsert ${block.name}.sql;\n"
@@ -175,9 +173,9 @@ class DbtManagerTest : FunSpec({
     }
 
     context("Should update a block while the block has some changes.") {
-        mockkObject(GitUtilsV2)
-        every { GitUtilsV2.checkoutMasterAndPull(any(), any()) } returns mockk()
-        every { GitUtilsV2.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
+        mockkObject(GitUtil)
+        every { GitUtil.checkoutMasterAndPull(any(), any()) } returns mockk()
+        every { GitUtil.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
 
         every { pm.getProfileById(workspaceId) } returns
                 mapper.readValue<ProfileEntity>(getResourceFileURL("/profiles/profile_bigquery.json"))
@@ -186,7 +184,7 @@ class DbtManagerTest : FunSpec({
         // mock env and data
         val expireBlock = providesQuestionBlockContent("/dbt/question_block_expire.json")
         val latestBlock = providesQuestionBlockContent("/dbt/question_block_latest.json")
-        val dbtManager = DbtManagerV2(pm)
+        val dbtManager = DbtManager(pm)
         dbtManager.reloadContext()
         val telleryDir = dbtManager.getContext()?.repoDir?.resolve("models/tellery")!!
         FileUtils.forceMkdir(telleryDir.toFile())
@@ -200,7 +198,7 @@ class DbtManagerTest : FunSpec({
         telleryDir.listDirectoryEntries().map { it.name } shouldContain "${latestBlock.name}.sql"
         blockPath.readText() shouldBe latestBlock.sql.toString()
         verify(exactly = 1) {
-            GitUtilsV2.checkoutNewBranchAndCommitAndPush(
+            GitUtil.checkoutNewBranchAndCommitAndPush(
                 any(),
                 any(),
                 "feat(tellery): async model files in tellery dir.\nupdate ${latestBlock.name}.sql;\n"
@@ -209,9 +207,9 @@ class DbtManagerTest : FunSpec({
     }
 
     context("Should delete the block while the block isn't exist.") {
-        mockkObject(GitUtilsV2)
-        every { GitUtilsV2.checkoutMasterAndPull(any(), any()) } returns mockk()
-        every { GitUtilsV2.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
+        mockkObject(GitUtil)
+        every { GitUtil.checkoutMasterAndPull(any(), any()) } returns mockk()
+        every { GitUtil.checkoutNewBranchAndCommitAndPush(any(), any(), any()) } returns mockk()
 
         every { pm.getProfileById(workspaceId) } returns
                 mapper.readValue<ProfileEntity>(getResourceFileURL("/profiles/profile_bigquery.json"))
@@ -219,7 +217,7 @@ class DbtManagerTest : FunSpec({
 
         // mock env and data
         val block = providesQuestionBlockContent("/dbt/question_block_latest.json")
-        val dbtManager = DbtManagerV2(pm)
+        val dbtManager = DbtManager(pm)
         dbtManager.reloadContext()
         val telleryDir = dbtManager.getContext()?.repoDir?.resolve("models/tellery")!!
         FileUtils.forceMkdir(telleryDir.toFile())
@@ -232,7 +230,7 @@ class DbtManagerTest : FunSpec({
         // assert
         telleryDir.listDirectoryEntries().map { it.name } shouldNotContain "${block.name}.sql"
         verify(exactly = 1) {
-            GitUtilsV2.checkoutNewBranchAndCommitAndPush(
+            GitUtil.checkoutNewBranchAndCommitAndPush(
                 any(),
                 any(),
                 "feat(tellery): async model files in tellery dir.\ndelete ${block.name}.sql;\n"
@@ -246,7 +244,7 @@ class DbtManagerTest : FunSpec({
         every { pm.getIntegrationInProfileAndByType(workspaceId, "dbt") } returns integrationEntity
 
         // mock env and data
-        val dbtManager = DbtManagerV2(pm)
+        val dbtManager = DbtManager(pm)
         dbtManager.reloadContext()
         val manifestPath = dbtManager.getContext()?.repoDir?.resolve("target/manifest.json")!!
         val mockManifestString = getResourceString("/dbt/manifest.json")
