+++ packages/connector/server/test/managers/FileProfileManagerTest.kt
@@ -1,4 +1,4 @@
-package managers
+package io.tellery.managers
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
