+++ packages/protobufs/base.proto
@@ -1,0 +1,11 @@
+syntax = "proto3";
+
+package tellery.grpc;
+option java_package = "io.tellery.base";
+option java_outer_classname = "TelleryBaseProto";
+option java_multiple_files = true;
+
+message KVEntry {
+  string key = 1;
+  string value = 2;
+}
