+++ packages/protobufs/profile.proto
@@ -7,8 +7,8 @@ option java_multiple_files = true;
 
 import "google/protobuf/wrappers.proto";
 import "google/protobuf/empty.proto";
-import "connector.proto";
 import "config.proto";
+import "base.proto";
 
 service ProfileService {
   rpc GetProfileConfigs (google.protobuf.Empty) returns (AvailableConfigs);
